refactor:  드래그 선택 hook에 종료 이후 동작 추가 : 드래그 종료 동작 handler 제거
