+++ src/pages/user/Apply/components/ApplicationForm/InterviewScheduleSelector.tsx
@@ -1,29 +1,17 @@
 import { useDragSelection } from '@/pages/user/Apply/hook/useDragSelection';
 import { useUpdateFormValue } from '@/pages/user/Apply/hook/useUpdateFormData';
-import {
-  convertSelectionToTimeInterval,
-  generateTimeIntervalArray,
-  mergeContinuousTimeInterval,
-} from '@/pages/user/Apply/utils/time';
+import { generateTimeIntervalArray } from '@/pages/user/Apply/utils/time';
 import { Text } from '@/shared/components/Text';
 import { TimeSpan, Wrapper, DateText } from './index.styled';
 import type { InterviewSchedule } from '@/pages/user/Apply/type/apply';
 
 export const InterviewScheduleSelector = ({ availableTime, date }: InterviewSchedule) => {
-  const { updateScheduleData } = useUpdateFormValue();
   const timeIntervalArray: [string, string][] = generateTimeIntervalArray(availableTime);
-
-  const handleDragEnd = () => {
-    const selectedInterviewTime: Set<string> = convertSelectionToTimeInterval(
-      selectedTime,
-      timeIntervalArray,
-    );
-    const mergedInterviewTime: string[] = mergeContinuousTimeInterval(selectedInterviewTime);
-    updateScheduleData(date, mergedInterviewTime);
-  };
+  const { updateScheduleData } = useUpdateFormValue();
 
   const { handleMouseDown, handleMouseMove, handleMouseUp, selectedTime } = useDragSelection(
-    handleDragEnd,
+    updateScheduleData,
+    date,
     timeIntervalArray,
   );
